fixed hyperparams
base_lr and end_lr have been removed in favor of base_frac and end_frac. Previously base_lr or end_lr could be greater than peak_lr. That is not the intention when we know that warmup-peak and then cosine-annealing is pretty much pareto-optimal. the range for B1 was also reduced because it was extremely noisy. Leaderboard.md is blank but we will be publishing a leaderboard of our best model results for each dataset that we explore here